added filter function

+++ J03/weekTwo/PhoneBook/Person.java
@@ -36,7 +36,7 @@ public class Person  {
         return this.phoneNumber;
     }
 
-    public void setPhoneNumber(int phoneNumber) {
+    public void setPhoneNumber(long phoneNumber) {
         this.phoneNumber = phoneNumber;
     }
